Re-architecture integration tests.
