--- tests/authorized/payloads/did-metadata.json
@@ -1,45 +1,0 @@
-{
-    "@context": "https://w3id.org/did-resolution/v1",
-    "dereferencingMetadata": {
-        "contentType": "application/did+ld+json",
-        "retrieved": "2023-08-24T07:05:05Z",
-        "did": {
-            "didString": "did:cheqd:testnet:c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0",
-            "methodSpecificId": "c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0",
-            "method": "cheqd"
-        }
-    },
-    "contentStream": {
-        "created": "2023-01-25T11:58:10Z",
-        "versionId": "e5615fc2-6f13-42b1-989c-49576a574cef",
-        "linkedResourceMetadata": [
-            {
-                "resourceURI": "did:cheqd:testnet:c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0/resources/9ba3922e-d5f5-4f53-b265-fc0d4e988c77",
-                "resourceCollectionId": "c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0",
-                "resourceId": "9ba3922e-d5f5-4f53-b265-fc0d4e988c77",
-                "resourceName": "Demo Resource",
-                "resourceType": "String",
-                "mediaType": "application/json",
-                "resourceVersion": "",
-                "created": "2023-01-25T12:08:39Z",
-                "checksum": "e1dbc03b50bdb995961dc8843df6539b79d03bf49787ed6462189ee97d27eaf3",
-                "previousVersionId": null,
-                "nextVersionId": null
-            },
-            {
-                "resourceURI": "did:cheqd:testnet:c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0/resources/e733ebb7-c8dd-41ed-9d42-33bceea70952",
-                "resourceCollectionId": "c1685ca0-1f5b-439c-8eb8-5c0e85ab7cd0",
-                "resourceId": "e733ebb7-c8dd-41ed-9d42-33bceea70952",
-                "resourceName": "ResourceName",
-                "resourceType": "String",
-                "mediaType": "text/plain; charset=utf-8",
-                "resourceVersion": "",
-                "created": "2023-01-25T12:04:52Z",
-                "checksum": "cffd829b06797f85407be9353056db722ca3eca0c05ab0462a42d30f19cdef09",
-                "previousVersionId": null,
-                "nextVersionId": null
-            }
-        ]
-    },
-    "contentMetadata": {}
-}
