+++ tests/constants.ts
@@ -122,3 +122,11 @@ export const INVALID_CREDENTIAL = {
 		jwt: "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJ2YyI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vc2NoZW1hLm9yZyIsImh0dHBzOi8vdmVyYW1vLmlvL2NvbnRleHRzL3Byb2ZpbGUvdjEiLCJodHRwczovL3czaWQub3JnL3ZjLXN0YXR1cy1saXN0LTIwMjEvdjEiXSwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlBlcnNvbiJdLCJjcmVkZW50aWFsU3ViamVjdCI6eyJnZW5kZXIiOiJtYWxlIiwibmFtZSI6IkJvYiJ9LCJjcmVkZW50aWFsU3RhdHVzIjp7ImlkIjoiaHR0cHM6Ly9yZXNvbHZlci5jaGVxZC5uZXQvMS4wL2lkZW50aWZpZXJzL2RpZDpjaGVxZDp0ZXN0bmV0OmQ0YTEzMDAzLTBiYzUtNDYwOC1iMjNhLTU0ZWE5MGZlOWY5MD9yZXNvdXJjZU5hbWU9Y2hlcWQtZW1wbG95ZWUtY3JlZGVudGlhbHMmcmVzb3VyY2VUeXBlPVN0YXR1c0xpc3QyMDIxUmV2b2NhdGlvbiMxMCIsInR5cGUiOiJTdGF0dXNMaXN0MjAyMUVudHJ5Iiwic3RhdHVzUHVycG9zZSI6InJldm9jYXRpb24iLCJzdGF0dXNMaXN0SW5kZXgiOiIxMCJ9fSwic3ViIjoiZGlkOmtleTp6Nk1raGFYZ0JaRHZvdERrTDUyNTdmYWl6dGlHaUMyUXRLTEdwYm5uRUd0YTJkb0siLCJuYmYiOjE2OTQwODIyMDEsImlzcyI6ImRpZDpjaGVxZDp0ZXN0bmV0OmQ0YTEzMDAzLTBiYzUtNDYwOC1iMjNhLTU0ZWE5MGZlOWY5MCJ9.AsW39Jej_Qae2FbpVdzyUhhuTi9gVWTx9w5NEyUfJYM6gQuCSAigiwcmaZZrcBWjOm53NXi-jRP561BpdVuSBw"
 	}
 };
+
+const PAYLOAD_BASE_PATH = "./tests/payloads";
+export enum PAYLOADS_PATH {
+	DID = `${PAYLOAD_BASE_PATH}/did`,
+	RESOURCE = `${PAYLOAD_BASE_PATH}/resource`,
+	CREDENTIAL = `${PAYLOAD_BASE_PATH}/credential`,
+	CREDENTIAL_STATUS = `${PAYLOAD_BASE_PATH}/credential-status`
+};
