+++ tests/credential-status/check.no-auth.spec.ts
@@ -1,5 +1,10 @@
+import {
+    TESTNET_DID,
+    NOT_EXISTENT_TESTNET_DID,
+    NOT_EXISTENT_STATUS_LIST_NAME,
+    PAYLOADS_PATH
+} from '../constants';
 import * as fs from 'fs';
-import { TESTNET_DID } from '../../constants';
 import { test, expect } from '@playwright/test';
 import { StatusCodes } from 'http-status-codes';
 
@@ -14,8 +19,8 @@ test('[Positive] It can check an unencrypted status-list with an existent body a
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(fs.readFileSync(
-        './tests/unauthorized/payloads/credential-status-check/status-revocation.json', 'utf-8')
+    const expected = JSON.parse(
+        fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL_STATUS}/check-status-revocation.json`, 'utf-8')
     );
 
     expect(body).toStrictEqual(expected);
@@ -32,8 +37,8 @@ test('Positive] It can check an unencrypted status-list with an existent body an
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(fs.readFileSync(
-        './tests/unauthorized/payloads/credential-status-check/status-suspension.json', 'utf-8')
+    const expected = JSON.parse(
+        fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL_STATUS}/check-status-suspension.json`, 'utf-8')
     );
 
     expect(body).toStrictEqual(expected);
@@ -52,7 +57,7 @@ test('Positive] It can check an unencrypted status-list with an existent body an
 
 //     const body = await response.json();
 //     const expected = JSON.parse(
-//         fs.readFileSync('./tests/unauthorized/payloads/credential-status-check/status-suspension.json', 'utf-8')
+//         fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL_STATUS}/check-status-suspension.json`, 'utf-8')
 //     );
 
 //     expect(body).toStrictEqual(expected);
@@ -72,8 +77,46 @@ test('Positive] It can check an unencrypted status-list with an existent body an
 
 //     const body = await response.json();
 //     const expected = JSON.parse(
-//         fs.readFileSync('./tests/unauthorized/payloads/credential-status-check/status-suspension.json', 'utf-8')
+//         fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL_STATUS}/check-status-suspension.json`, 'utf-8')
 //     );
 
 //     expect(body).toStrictEqual(expected);
 // });
+
+test('[Negative] It cannot check credential-status with not existent DID', async ({ request }) => {
+    const response = await request.post('/credential-status/check?statusPurpose=revocation', {
+        data: {
+            did: NOT_EXISTENT_TESTNET_DID,
+            index: 10,
+            statusListName: "cheqd-employee-credentials",
+        }
+    });
+    expect(response.status()).toBe(StatusCodes.NOT_FOUND);
+
+    const body = await response.json();
+    const expected = {
+        checked: false,
+        error: "check: error: status list 'cheqd-employee-credentials' not found"
+    };
+
+    expect(body).toStrictEqual(expected);
+});
+
+test('[Negative] It cannot check credential-status with an existent DID and not existent statusListName', async ({ request }) => {
+    const response = await request.post('/credential-status/check?statusPurpose=revocation', {
+        data: {
+            did: TESTNET_DID,
+            index: 10,
+            statusListName: NOT_EXISTENT_STATUS_LIST_NAME,
+        }
+    });
+    expect(response.status()).toBe(StatusCodes.NOT_FOUND);
+
+    const body = await response.json();
+    const expected = {
+        checked: false,
+        error: `check: error: status list '${NOT_EXISTENT_STATUS_LIST_NAME}' not found`
+    };
+
+    expect(body).toStrictEqual(expected);
+});
