+++ tests/credential/verify.no-auth.spec.ts
@@ -1,7 +1,15 @@
+import {
+    VALID_CREDENTIAL,
+    VALID_JWT_TOKEN,
+    INVALID_JWT_TOKEN,
+    // INVALID_CREDENTIAL,
+    NOT_MATCHED_CREDENTIAL_AND_JWT,
+    PAYLOADS_PATH
+} from '../constants';
 import * as fs from 'fs';
 import { test, expect } from '@playwright/test';
-import { VALID_CREDENTIAL, VALID_JWT_TOKEN } from '../../constants';
 import { StatusCodes } from 'http-status-codes';
+
 
 test('[Positive] It can verify credential with a valid JWT body', async ({ request }) => {
     const response = await request.post('/credential/verify', {
@@ -10,9 +18,7 @@ test('[Positive] It can verify credential with a valid JWT body', async ({ reque
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-jwt.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-jwt.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -24,9 +30,7 @@ test('[Positive] It can verify credential with a valid credential body', async (
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-credential.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-credential.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -39,9 +43,7 @@ test('[Positive] It can verify credential with a valid JWT body and verifyStatus
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(fs.readFileSync(
-        './tests/unauthorized/payloads/credential-verify/verify-jwt-status.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-jwt-status.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -53,9 +55,7 @@ test('[Positive] It can verify credential with a valid credential body and verif
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-credential-status.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-credential-status.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -68,9 +68,7 @@ test('[Positive] It can verify credential with a valid JWT body and fetchRemoteC
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-jwt.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-jwt.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -82,9 +80,7 @@ test('[Positive] It can verify credential with a valid credential body and fetch
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-credential.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-credential.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -96,9 +92,7 @@ test('[Positive] It can verify credential with a valid JWT body, verifyStatus=tr
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(fs.readFileSync(
-        './tests/unauthorized/payloads/credential-verify/verify-jwt-status.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-jwt-status.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
@@ -111,9 +105,57 @@ test('[Positive] It can verify credential with a valid credential body, verifySt
     expect(response.status()).toBe(StatusCodes.OK);
 
     const body = await response.json();
-    const expected = JSON.parse(
-        fs.readFileSync('./tests/unauthorized/payloads/credential-verify/verify-credential-status.json', 'utf-8')
-    );
+    const expected = JSON.parse(fs.readFileSync(`${PAYLOADS_PATH.CREDENTIAL}/verify-credential-status.json`, 'utf-8'));
 
     expect(body).toStrictEqual(expected);
 });
+
+test('[Negative] It cannot verify credential with an invalid credential body where credential and JWT are different', async ({ request }) => {
+    const response = await request.post('/credential/verify', {
+        data: { credential: NOT_MATCHED_CREDENTIAL_AND_JWT }
+    });
+    expect(response.status()).toBe(StatusCodes.BAD_REQUEST);
+
+    const body = await response.json();
+    const expected = {
+        verified: false,
+        error: {} // TODO: return an exact error instead of empty result
+    };
+
+    expect(body).toStrictEqual(expected);
+});
+
+test('[Negative] It cannot verify credential with an invalid JWT body', async ({ request }) => {
+    const response = await request.post('/credential/verify', {
+        data: { credential: INVALID_JWT_TOKEN }
+    });
+    expect(response.status()).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
+
+    const body = await response.json();
+    const expected = {
+        authenticated: false,
+        error: "InvalidTokenError: Invalid token specified: Cannot read properties of undefined (reading 'replace')",
+        customerId: null
+    };
+
+    expect(body).toStrictEqual(expected);
+});
+
+
+// This test returns 400 when user is authenticated, but in case of when user is not authenticated it returns 500
+// test('[Negative] It cannot verify credential with an invalid credential body', async ({ request }) => {
+//     const response = await request.post('/credential/verify', {
+//         data: { credential: INVALID_CREDENTIAL }
+//     });
+//     expect(response.status()).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
+
+//     const body = await response.json();
+//     const expected = {
+//         authenticated: false,
+//         error: "InvalidTokenError: Invalid token specified: Cannot read properties of undefined (reading 'replace')",
+//         customerId: null
+//     };
+
+//     expect(body).toStrictEqual(expected);
+// });
+
