+++ tests/did/search.no-auth.spec.ts
@@ -4,8 +4,10 @@ import {
     TESTNET_DID_FRAGMENT,
     TESTNET_DID_IDENTIFIER,
     APPLICATION_DID_LD_JSON,
-    TESTNET_DID_CREATED_TIME
-} from '../../constants';
+    TESTNET_DID_CREATED_TIME,
+    NOT_EXISTENT_TESTNET_DID,
+    NOT_EXISTENT_TESTNET_DID_IDENTIFIER
+} from '../constants';
 import { test, expect } from '@playwright/test';
 import { StatusCodes } from 'http-status-codes';
 
@@ -88,4 +90,30 @@ test('[Positive] It can search with an existent DID and fragment', async ({ requ
     expect(body.dereferencingMetadata.did).toStrictEqual(expected.dereferencingMetadata.did);
     expect(body.contentStream.id).toBe(expected.contentStream.id);
     expect(body.contentMetadata.created).toBe(expected.contentMetadata.created);
+});
+
+test('[Negative] It cannot search not existent DID', async ({ request }) => {
+    const response = await request.get(`/did/search/${NOT_EXISTENT_TESTNET_DID}`);
+    expect(response.status()).toBe(StatusCodes.NOT_FOUND);
+
+    const body = await response.json();
+    const expected = {
+        didResolutionMetadata: {
+            contentType: APPLICATION_DID_LD_JSON,
+            error: "notFound",
+            did: {
+                didString: NOT_EXISTENT_TESTNET_DID,
+                methodSpecificId: NOT_EXISTENT_TESTNET_DID_IDENTIFIER,
+                method: DID_METHOD
+            }
+        },
+        didDocument: null,
+        didDocumentMetadata: {}
+    };
+
+    expect(body.didResolutionMetadata.contentType).toBe(expected.didResolutionMetadata.contentType);
+    expect(body.didResolutionMetadata.error).toBe(expected.didResolutionMetadata.error);
+    expect(body.didResolutionMetadata.did).toStrictEqual(expected.didResolutionMetadata.did);
+    expect(body.didDocument).toBe(expected.didDocument);
+    expect(body.didDocumentMetadata).toStrictEqual(expected.didDocumentMetadata);
 });
