+++ tests/payloads/credential-status/create-encrypted-without-permissions.json
