+++ tests/payloads/credential-status/create-unencrypted-without-permissions.json
