+++ tests/payloads/credential-status/update-encrypted-without-permissions.json
