+++ tests/payloads/credential-status/update-unencrypted-without-permissions.json
