+++ tests/payloads/credential/revoke-credential-without-permissions.json
