+++ tests/payloads/credential/verify-jwt.json
