+++ tests/presentation/verify.no-auth.spec.ts
@@ -1,15 +1,22 @@
 import { test } from '@playwright/test';
 
-
 // TODO: Complete tests.
+test('[Positive] It can verify a valid JWT body', async () => {
+    ;
+});
+
+test('[Positive] It can verify a valid presentation body', async () => {
+    ;
+});
+
 test('[Negative] It cannot verify presentation with an invalid credential body, where credential and JWT are different', async () => {
     ;
-})
+});
 
 test('[Negative] It cannot verify presentation with an invalid JWT body', async () => {
     ;
-})
+});
 
 test('[Negative] It cannot verify presentation with an invalid credential body', async () => {
     ;
-})
+});
