--- tests/unauthorized/presentation-verify/positive.spec.ts
@@ -1,10 +1,0 @@
-import { test } from '@playwright/test';
-
-// TODO: Complete tests.
-test('[Positive] It can verify a valid JWT body', async () => {
-    ;
-})
-
-test('[Positive] It can verify a valid presentation body', async () => {
-    ;
-})
